+++ app/db/base.py
@@ -118,11 +118,11 @@ def init_models(
         "app.wms.stock.models.stock_snapshot",
         "app.wms.inbound.models.inbound_event",
         "app.wms.outbound.models.outbound_event",
-        "app.models.order",
-        "app.models.order_item",
-        "app.models.order_address",
-        "app.models.order_logistics",
-        "app.models.order_fulfillment",
+        "app.oms.orders.models.order",
+        "app.oms.orders.models.order_item",
+        "app.oms.orders.models.order_address",
+        "app.oms.orders.models.order_logistics",
+        "app.oms.orders.models.order_fulfillment",
         "app.models.store",
         "app.models.warehouse",
         "app.models.platform_shops",
@@ -147,6 +147,7 @@ def init_models(
         "app.tms.pricing.templates.models",
         "app.tms.providers.models",
         "app.tms.shipment.models",
+        "app.oms.orders.models",
     ):
         for mod in _iter_model_modules_recursive(pkg_name):
             if mod in ex or mod in loaded:
