+++ app/oms/orders/models/__init__.py
@@ -1,0 +1,20 @@
+# app/oms/orders/models/__init__.py
+# Domain-owned ORM models for OMS orders.
+
+from app.oms.orders.models.order import Order
+from app.oms.orders.models.order_address import OrderAddress
+from app.oms.orders.models.order_fulfillment import OrderFulfillment
+from app.oms.orders.models.order_item import OrderItem
+from app.oms.orders.models.order_line import OrderLine
+from app.oms.orders.models.order_logistics import OrderLogistics
+from app.oms.orders.models.order_state_snapshot import OrderStateSnapshot
+
+__all__ = [
+    "Order",
+    "OrderAddress",
+    "OrderFulfillment",
+    "OrderItem",
+    "OrderLine",
+    "OrderLogistics",
+    "OrderStateSnapshot",
+]
