+++ app/oms/platforms/models/pdd_order_order_mapping.py
@@ -11,7 +11,7 @@ from sqlalchemy.orm import Mapped, mapped_column, relationship
 from app.db.base import Base
 
 if TYPE_CHECKING:
-    from app.models.order import Order
+    from app.oms.orders.models.order import Order
     from app.oms.platforms.models.pdd_order import PddOrder
